+++ include/ozz/animation/runtime/animation.h
@@ -64,6 +64,14 @@ class OZZ_ANIMATION_DLL Animation {
   // Builds a default animation.
   Animation();
 
+  // Allow moves.
+  Animation(Animation&&);
+  Animation& operator=(Animation&&);
+
+  // Delete copies.
+  Animation(Animation const&) = delete;
+  Animation& operator=(Animation const&) = delete;
+
   // Declares the public non-virtual destructor.
   ~Animation();
 
@@ -81,9 +89,7 @@ class OZZ_ANIMATION_DLL Animation {
   const char* name() const { return name_ ? name_ : ""; }
 
   // Gets the buffer of translations keys.
-  span<const Float3Key> translations() const {
-    return translations_;
-  }
+  span<const Float3Key> translations() const { return translations_; }
 
   // Gets the buffer of rotation keys.
   span<const QuaternionKey> rotations() const { return rotations_; }
@@ -99,12 +105,7 @@ class OZZ_ANIMATION_DLL Animation {
   void Save(ozz::io::OArchive& _archive) const;
   void Load(ozz::io::IArchive& _archive, uint32_t _version);
 
- protected:
  private:
-  // Disables copy and assignation.
-  Animation(Animation const&);
-  void operator=(Animation const&);
-
   // AnimationBuilder class is allowed to instantiate an Animation.
   friend class offline::AnimationBuilder;